Introduce BannerExecutor
